# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "askgpt", "description": "distributed gpt client built with love", "version": "1"}, "paths": {"/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_auth-login"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/auth/signup": {"post": {"tags": ["auth"], "summary": "Signup User", "description": "Request will be redirected to user route for user info", "operationId": "signup_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/": {"get": {"tags": ["user"], "summary": "Find User By Email", "operationId": "find_user_by_email", "parameters": [{"name": "email", "in": "query", "required": true, "schema": {"type": "string", "title": "Email"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/PublicUserInfo"}, {"type": "null"}], "title": "Response User-Find User By Email"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/me": {"get": {"tags": ["user"], "summary": "Get Public User", "operationId": "get_public_user", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/v1/users/{user_id}": {"get": {"tags": ["user"], "summary": "Get User Detail", "description": "Private user info", "operationId": "get_user_detail", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/UserAuth"}, {"type": "null"}], "title": "Response User-Get User Detail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["user"], "summary": "Delete User", "operationId": "delete_user", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/apikeys": {"post": {"tags": ["user"], "summary": "Add Api Key", "operationId": "add_api_key", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserAddAPIRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/v1/gpt/openai/sessions": {"get": {"tags": ["gpt"], "summary": "List Sessions", "operationId": "list_sessions", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/PulibcSessionInfo"}, "type": "array", "title": "Response Gpt-List Sessions"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["gpt"], "summary": "Create Session", "operationId": "create_session", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/v1/gpt/openai/sessions/{session_id}": {"get": {"tags": ["gpt"], "summary": "Get Session", "operationId": "get_session", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["gpt"], "summary": "Rename Session", "operationId": "rename_session", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionRenameRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["gpt"], "summary": "Delete Session", "operationId": "delete_session", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/gpt/openai/chat/{session_id}": {"post": {"tags": ["gpt"], "summary": "Chat", "operationId": "chat", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatCompletionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/health": {"get": {"tags": ["health check"], "summary": "<Lambda>", "operationId": "<lambda>", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_auth-login": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_auth-login"}, "ChatCompletionRequest": {"properties": {"question": {"type": "string", "title": "Question"}, "role": {"type": "string", "enum": ["system", "user", "assistant", "function"], "title": "Role"}, "model": {"type": "string", "enum": ["gpt-3.5-turbo", "gpt-3.5-turbo-16k", "gpt-4", "gpt-4-32k", "gpt-4-1106-preview", "gpt-4-vision-preview"], "title": "Model", "default": "gpt-3.5-turbo"}, "frequency_penalty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Frequency Penalty"}, "logit_bias": {"anyOf": [{"additionalProperties": {"type": "integer"}, "type": "object"}, {"type": "null"}], "title": "Logit Bias"}, "max_tokens": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Max Tokens"}, "n": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "N"}, "presence_penalty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Presence Penalty"}, "response_format": {"title": "Response Format"}, "seed": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Seed"}, "stop": {"anyOf": [{"type": "string"}, {"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Stop"}, "stream": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Stream"}, "temperature": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Temperature"}, "tool_choice": {"title": "Tool Choice"}, "tools": {"anyOf": [{"items": {}, "type": "array"}, {"type": "null"}], "title": "Tools"}, "top_p": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Top P"}, "user": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "User"}, "extra_headers": {"anyOf": [{"additionalProperties": {"anyOf": [{"type": "string"}, {"const": false}]}, "type": "object"}, {"type": "null"}], "title": "Extra Headers", "default": {}}, "extra_query": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Extra Query", "default": {}}, "extra_body": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Extra Body", "default": {}}, "timeout": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Timeout", "default": 120}}, "additionalProperties": false, "type": "object", "required": ["question", "role"], "title": "ChatCompletionRequest", "description": "    Creates a model response for the given chat conversation.\n\nArgs:\n  messages: A list of messages comprising the conversation so far.\n      [Example Python code](https://cookbook.openai.com/examples/how_to_format_inputs_to_chatgpt_models).\n\n  model: ID of the model to use. See the\n      [model endpoint compatibility](https://platform.openai.com/docs/models/model-endpoint-compatibility)\n      table for details on which models work with the Chat API.\n\n  frequency_penalty: Number between -2.0 and 2.0. Positive values penalize new tokens based on their\n      existing frequency in the text so far, decreasing the model's likelihood to\n      repeat the same line verbatim.\n\n      [See more information about frequency and presence penalties.](https://platform.openai.com/docs/guides/gpt/parameter-details)\n\n\n  logit_bias: Modify the likelihood of specified tokens appearing in the completion.\n\n      Accepts a JSON object that maps tokens (specified by their token ID in the\n      tokenizer) to an associated bias value from -100 to 100. Mathematically, the\n      bias is added to the logits generated by the model prior to sampling. The exact\n      effect will vary per model, but values between -1 and 1 should decrease or\n      increase likelihood of selection; values like -100 or 100 should result in a ban\n      or exclusive selection of the relevant token.\n\n  max_tokens: The maximum number of [tokens](/tokenizer) to generate in the chat completion.\n\n      The total length of input tokens and generated tokens is limited by the model's\n      context length.\n      [Example Python code](https://cookbook.openai.com/examples/how_to_count_tokens_with_tiktoken)\n      for counting tokens.\n\n  n: How many chat completion choices to generate for each input message.\n\n  presence_penalty: Number between -2.0 and 2.0. Positive values penalize new tokens based on\n      whether they appear in the text so far, increasing the model's likelihood to\n      talk about new topics.\n\n      [See more information about frequency and presence penalties.](https://platform.openai.com/docs/guides/gpt/parameter-details)\n\n  response_format: An object specifying the format that the model must output. Used to enable JSON\n      mode.\n\n  seed: This feature is in Beta. If specified, our system will make a best effort to\n      sample deterministically, such that repeated requests with the same `seed` and\n      parameters should return the same result. Determinism is not guaranteed, and you\n      should refer to the `system_fingerprint` response parameter to monitor changes\n      in the backend.\n\n  stop: Up to 4 sequences where the API will stop generating further tokens.\n\n  stream: If set, partial message deltas will be sent, like in ChatGPT. Tokens will be\n      sent as data-only\n      [server-sent events](https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events#Event_stream_format)\n      as they become available, with the stream terminated by a `data: [DONE]`\n      message.\n      [Example Python code](https://cookbook.openai.com/examples/how_to_stream_completions).\n\n  temperature: What sampling temperature to use, between 0 and 2. Higher values like 0.8 will\n      make the output more random, while lower values like 0.2 will make it more\n      focused and deterministic.\n\n      We generally recommend altering this or `top_p` but not both.\n\n  tool_choice: Controls which (if any) function is called by the model. `none` means the model\n      will not call a function and instead generates a message. `auto` means the model\n      can pick between generating a message or calling a function. Specifying a\n      particular function via\n      `{\"type: \"function\", \"function\": {\"name\": \"my_function\"}}` forces the model to\n      call that function.\n\n      `none` is the default when no functions are present. `auto` is the default if\n      functions are present.\n\n  tools: A list of tools the model may call. Currently, only functions are supported as a\n      tool. Use this to provide a list of functions the model may generate JSON inputs\n      for.\n\n  top_p: An alternative to sampling with temperature, called nucleus sampling, where the\n      model considers the results of the tokens with top_p probability mass. So 0.1\n      means only the tokens comprising the top 10% probability mass are considered.\n\n      We generally recommend altering this or `temperature` but not both.\n\n  user: A unique identifier representing your end-user, which can help OpenAI to monitor\n      and detect abuse.\n      [Learn more](https://platform.openai.com/docs/guides/safety-best-practices/end-user-ids).\n\n  extra_headers: Send extra headers\n\n  extra_query: Add additional query parameters to the request\n\n  extra_body: Add additional JSON properties to the request\n\n  timeout: Override the client-level default timeout for this request, in seconds", "examples": [{"model": "gpt-4-1106-preview", "question": "enter your question here", "role": "user", "stream": true}]}, "CreateUserRequest": {"properties": {"user_name": {"type": "string", "title": "User Name", "default": ""}, "email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "additionalProperties": false, "type": "object", "required": ["email", "password"], "title": "CreateUserRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "PublicUserInfo": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "user_name": {"type": "string", "title": "User Name"}, "email": {"type": "string", "title": "Email"}}, "type": "object", "required": ["user_id", "user_name", "email"], "title": "PublicUserInfo"}, "PulibcSessionInfo": {"properties": {"session_id": {"type": "string", "title": "Session Id"}, "session_name": {"type": "string", "title": "Session Name"}}, "type": "object", "required": ["session_id", "session_name"], "title": "PulibcSessionInfo"}, "SessionRenameRequest": {"properties": {"name": {"type": "string", "title": "Name"}}, "additionalProperties": false, "type": "object", "required": ["name"], "title": "SessionRenameRequest"}, "SupportedGPTs": {"const": "openai"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "TokenResponse"}, "UserAddAPIRequest": {"properties": {"api_key": {"type": "string", "title": "Api Key"}, "api_type": {"allOf": [{"$ref": "#/components/schemas/SupportedGPTs"}], "default": "openai"}}, "additionalProperties": false, "type": "object", "required": ["api_key"], "title": "UserAddAPIRequest"}, "UserAuth": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "role": {"allOf": [{"$ref": "#/components/schemas/UserRoles"}], "default": "user"}, "credential": {"$ref": "#/components/schemas/UserCredential"}, "last_login": {"type": "string", "format": "date-time", "title": "Last Login"}, "is_active": {"type": "boolean", "title": "Is Active", "default": true}}, "type": "object", "required": ["user_id", "credential", "last_login"], "title": "UserAuth"}, "UserCredential": {"properties": {"user_name": {"type": "string", "title": "User Name", "default": ""}, "user_email": {"type": "string", "format": "email", "title": "User Email"}, "hash_password": {"type": "string", "title": "Hash Password"}}, "type": "object", "required": ["user_email", "hash_password"], "title": "UserCredential"}, "UserRoles": {"type": "string", "enum": ["admin", "user"], "title": "UserRoles"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "auth/login"}}}}}}